+++ package.json
@@ -25,6 +25,7 @@
     "react-dom": "18.0.0",
     "react-native": "0.69.5",
     "react-native-background-timer": "^2.4.1",
+    "react-native-elements": "^3.4.2",
     "react-native-safe-area-context": "4.3.1",
     "react-native-screens": "~3.15.0",
     "react-native-web": "~0.18.7",
