+++ src/screens/TimerScreen.tsx
@@ -1,5 +1,4 @@
 import { useTailwind } from "tailwind-rn/dist";
-import React, { useState, useEffect } from "react";
 import {
   View,
   Text,
@@ -8,81 +7,25 @@ import {
   Dimensions,
   StyleSheet,
 } from "react-native";
+import Timer from "@components/Timer";
+import Category from "@components/Category";
 import type { NativeStackScreenProps } from "@react-navigation/native-stack";
 import { RootStackParamList } from "../../App";
-import { StatusBar } from "expo-status-bar";
 
-// type Props = NativeStackScreenProps<RootStackParamList, "Timer">;
+type Props = NativeStackScreenProps<RootStackParamList, "Timer">;
 
-const screen = Dimensions.get("window");
-const formatNumber = (number) => `0${number}`.slice(-2);
-
-const getRemaining = (time) => {
-  const mins = Math.floor(time / 60);
-  const secs = time - mins * 60;
-  return { mins: formatNumber(mins), secs: formatNumber(secs) };
-};
 const TimerScreen = () => {
   const tailwind = useTailwind();
-  const [remainingSecs, setRemainingSecs] = useState(0);
-  const [isActive, setIsActive] = useState(false);
-  const { mins, secs } = getRemaining(remainingSecs);
-
-  const toggle = () => {
-    setIsActive(!isActive);
-  };
-
-  useEffect(() => {
-    let interval = null;
-    if (isActive) {
-      interval = setInterval(() => {
-        setRemainingSecs((remainingSecs) => remainingSecs + 1);
-      }, 1000);
-    } else if (!isActive && remainingSecs !== 0) {
-      clearInterval(interval);
-    }
-    return () => clearInterval(interval);
-  }, [isActive, remainingSecs]);
   return (
     <SafeAreaView style={tailwind("flex-1")}>
       <View style={tailwind("flex flex-col items-start")}>
         <Text>Timerデータを表示します</Text>
       </View>
-      <View style={styles.container}>
-        <StatusBar style="light-content" />
-        <Text style={styles.timerText}>{`${mins}:${secs}`}</Text>
-        <TouchableOpacity onPress={toggle} style={styles.button}>
-          <Text style={styles.buttonText}>{isActive ? "Pause" : "Start"}</Text>
-        </TouchableOpacity>
+      <View>
+        <Category />
       </View>
+      <Timer />
     </SafeAreaView>
   );
 };
 export default TimerScreen;
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: "#071218",
-    alignItems: "center",
-    justifyContent: "center",
-  },
-  button: {
-    borderWidth: 10,
-    borderColor: "#B9AAFF",
-    width: screen.width / 2,
-    height: screen.width / 2,
-    borderRadius: screen.width / 2,
-    alignItems: "center",
-    justifyContent: "center",
-  },
-  buttonText: {
-    fontSize: 45,
-    color: "#B9AAFF",
-  },
-  timerText: {
-    color: "#fff",
-    fontSize: 90,
-    marginBottom: 20,
-  },
-});
