+++ tailwind.css
@@ -7,20 +7,14 @@
   margin-bottom: 1rem
 }
 
-.mt-20 {
-  margin-top: 5rem
+.my-8 {
+  margin-top: 2rem;
+  margin-bottom: 2rem
 }
 
-.ml-12 {
-  margin-left: 3rem
-}
-
-.ml-1 {
-  margin-left: 0.25rem
-}
-
-.mt-4 {
-  margin-top: 1rem
+.my-14 {
+  margin-top: 3.5rem;
+  margin-bottom: 3.5rem
 }
 
 .mr-4 {
@@ -35,40 +29,20 @@
   display: flex
 }
 
-.h-2\/4 {
-  height: 50%
+.h-8 {
+  height: 2rem
 }
 
 .h-full {
   height: 100%
 }
 
-.h-8 {
-  height: 2rem
-}
-
-.h-10 {
-  height: 2.5rem
-}
-
-.w-3\/4 {
-  width: 75%
-}
-
-.w-40 {
-  width: 10rem
+.w-full {
+  width: 100%
 }
 
 .w-80 {
   width: 20rem
-}
-
-.w-60 {
-  width: 15rem
-}
-
-.w-full {
-  width: 100%
 }
 
 .flex-1 {
@@ -112,14 +86,9 @@
   border-width: 1px
 }
 
-.bg-stone-200 {
+.bg-black {
   --tw-bg-opacity: 1;
-  background-color: rgb(231 229 228 / var(--tw-bg-opacity))
-}
-
-.bg-white {
-  --tw-bg-opacity: 1;
-  background-color: rgb(255 255 255 / var(--tw-bg-opacity))
+  background-color: rgb(0 0 0 / var(--tw-bg-opacity))
 }
 
 .bg-violet-800 {
@@ -127,17 +96,22 @@
   background-color: rgb(91 33 182 / var(--tw-bg-opacity))
 }
 
+.bg-white {
+  --tw-bg-opacity: 1;
+  background-color: rgb(255 255 255 / var(--tw-bg-opacity))
+}
+
 .bg-sky-400 {
   --tw-bg-opacity: 1;
   background-color: rgb(56 189 248 / var(--tw-bg-opacity))
 }
 
-.p-4 {
-  padding: 1rem
-}
-
 .p-2 {
   padding: 0.5rem
+}
+
+.p-4 {
+  padding: 1rem
 }
 
 .py-4 {
@@ -145,20 +119,14 @@
   padding-bottom: 1rem
 }
 
-.pt-2 {
-  padding-top: 0.5rem
+.py-14 {
+  padding-top: 3.5rem;
+  padding-bottom: 3.5rem
 }
 
-.pl-2 {
-  padding-left: 0.5rem
-}
-
-.pl-1 {
-  padding-left: 0.25rem
-}
-
-.pt-4 {
-  padding-top: 1rem
+.py-10 {
+  padding-top: 2.5rem;
+  padding-bottom: 2.5rem
 }
 
 .text-2xl {
@@ -173,11 +141,6 @@
 
 .font-bold {
   font-weight: 700
-}
-
-.text-blue-900 {
-  --tw-text-opacity: 1;
-  color: rgb(30 58 138 / var(--tw-text-opacity))
 }
 
 .text-white {
